[kcm] Remove usage of tabs
The usage of tabs comes with quite some visual and technical problems. In general we tend to move away from tabs in KCMs for that reason.

The Device and Applications lists are combined into a single list that is shown as the initial view. The former Advanced tab is opened as a new page when clicking a Configure button. This is a pattern we use in other KCMs as well.

The content of the lists and the advanced tab are unchanged for now, but this opens the door for future further improvements