WIP: Introduce VolumeMonitor object to show volume of active streams/sinks/outputs
Several volume control UIs show a meter of which streams are actively
playing audio.

VolumeObject gains a property to list the a source that can be
monitored, and a QML exposed class exists to expose this in a very QML
friendly consumable way.

As we are technically creating a recording stream, I had to filter this
out so we don't show it again. I feel this should be resolvable in a
better way?