Show current Volume in Plasmoid sliders
We draw the background of slider now manually. The length of the highlight
shows the current volume. Minimum is the minimal width of the element,
maximum is highlight up to the handle position and when nothing is playing
no highlight is shown. An animation makes the transitions a bit smoother.
BUG:411563
BUG:428968