Show current Volume in Plasmoid sliders
We inlay a second slider highlight positioned like the one in the plasma style
with a different color representing the current volume.
Because we can't know which color a slider is using I am simply shifting the
highlightColor by 30 degree with full saturation to get a color that looks
similiar but should be still visible enough.