applet: Fix crash when opening context menu with no contents
Do not try to open the context menu via the right mouse button
or the Menu key press when it has no contents, because that
leads to a crash.

The menu button was already guarded against this because it
gets hidden in that case.

BUG: 470847
FIXED-IN: 5.27.6

(cherry picked from commit 65936f7a078f9cd83ef4c702e677b69db150f6cd)