allow user to rename devices and select a preferred name source
# Name Source

Allows choosing the name out of a set number of sources. Names have been
unified across applet and kcm now.
Changes here are instantly applied throughout the stack.

# Overrides

Additionally, overrides may be set. These override all supported name
fields so it doesn't matter which name source is chosen at the time the
change is made. The override installs a wireplumber json file and
restarts the stack to apply the change.