applet: warn user on raising max volume, and periodically reset it
This feature can be actively dangerous to your ears or your device's
speakers if you use it to raise the volume of sound being played at a
normal level (ask me how I know…).

However removing it entirely would remove the ability to boost
annoyingly quiet sounds, which is useful.

Boosting quiet volume on demand only when safe would be amazing, but is
not technically feasibly at this time.

So as a compromise between these constraints, let's warn the user about
the risks and turn the feature off on restart and sleep.