Dont show osd if the kcm is open or an expanded applet is shown
In the past the osd was not shown the user dragged a slider as it
makes little sense since the information is already on screen and
the user is interacting with this very control. This changed
with the daemon handling shortcuts and osd as it reacts to every
volume change.
This change brings that behavior back and makes it even a bit
better. If an applet is expanded from the systray, exists on the
desktop and the desktop focused or a kcm exists we dont show the osd.
For example, now when you scroll on the systray icon while the
applet or the kcm is open no osd is shown.