kded: Separate muting actions and improve usage of mute key
With the global mute, muting the primary device individually through e.g.
the applet would render it uncontrollable through usage of the mute key.
A user could keep pressing the mute key, but the device will remain muted.
I thought this was a little bit unintuitive and possibly frustrating.

In this MR, I split the actions into a global mute action, which always
disables/enables the global mute, and an action for the mute keys, which
will try to try to unmute the primary device along with the global muted
devices if mute is pressed when global mute is enabled, and unmute the
primary device when global mute is disabled but the primary device is
muted.