src/preferreddevice: Always prefer the default device
Otherwise, a new default device which is suspended (like bluetooth earphones) are unexpectedly ignored.

The class is changed from a more heavy construction around the default sink/source in pulseaudio-qt to just a simple wrapper. It is still necessary because QML also wants to know the preferred device, otherwise it could be removed entirely.